Tag Type Enum Flattening
Tried to write an object to a buffer with `writeUncompressed()`, but it wasn't being allowed as a valid object, because the types of the tag key types are typed as enum members, rather than plain strings. This allows you to use both, you can either pass in the enum to the built object, or a plain object with string keys itself can now be valid as well.

https://stackoverflow.com/questions/67863114/is-there-a-way-to-have-a-switch-statement-for-an-object-with-cases-that-validate

```ts
import * as P from "prismarine-nbt";

const nbt = { /*...*/ } satisfies P.NBT;
console.log(nbt);

P.writeUncompressed(nbt);
                 // ^? This was erroring
```